LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/evaluation-error-1-expected.txt

CONSOLE MESSAGE: line 2: [object Object]

FAIL Test that exceptions during evaluation lead to error events on window, and that exceptions are remembered.
 assert_array_equals: lengths differ, expected 5 got 2
